Create updated.html
In this version, an <audio> element is added to the page with the ID audioPlayer. The src attribute of the audio element is set to the URL of your audio file (song.mp3). The togglePlayPause() function now controls the playback of the audio element - it plays the audio when the play button is clicked and pauses it when clicked again.